compilation warning about signed/unsigned comparison fixed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@14816 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/menuitem.cpp
@@ -283,7 +283,7 @@ void wxMenuItem::Check(bool check)
     {
         if ( ::CheckMenuItem(hmenu,
                              GetRealId(),
-                             MF_BYCOMMAND | flags) == -1 )
+                             MF_BYCOMMAND | flags) == (DWORD)-1 )
         {
             wxLogLastError(wxT("CheckMenuItem"));
         }
